requirements.txt
mammoth==1.9.0
beautifulsoup4==4.12.3
pillow==12.1.0

